Introduce IdentifiedResource, bundling a resource and specification.
This moves Specifications to the resource level rather than registry
level as they should be.

TBD on whether we want/need to have multiple specifications on the same
resource, so far not, though in the test suite we kind of have this.

But with this... all tests pass (except for 1 draft 2019 test, the
invalid $defs test validating against the metaschema, which I'm
suspicious is either wrong or else is a bug in the recursiveRef code
rather than here).

Next step will be to move the just-referenced recursiveRef code here and
check.