Short-circuit when combining a registry with itself.
This is useful when combining some base registry with further
resources -- e.g. in jsonschema we do JSONSCHEMA_SPECS.combine(...)
and if no additional resources are provided that's meant to be a
no-op.

Doing this seems to be a huge speed-up, taking the test run from ~25s
to 9s (though even that is very slow compared to the non-compliant
implementation, so some performance optimization will come regardless).

But as a micro-optimization, this seems definitely useful.