Better handle traversing into subresources while resolving pointers.
This is .. ugly and duplicative. But more correct.

When traversing a JSON Pointer, we need to know when and if we're
entering a subresource, and this differs by spec (and has to do
with whether we've just entered a known keyword).

This will likely get less duplicative when #24 is done.