Don't convert pmaps to pmaps on initialization of Registries.
I.e. if it be a pmap, don't be pmapping it -- somewhat surprising
pmap doesn't already short-circuit here, but I have vague memories
of noticing this before.

This gives a ~20% speedup on a microbenchmark (of validator creation
in jsonschema which creates new Registries each time) and a ~10%
speedup on a jsonschema test run.

Thanks to @cfbolz for helping to discover it (though more problems
are lurking somewhere to find...).