Strip empty fragments when directly looking up a resource.
This isn't necessarily 100% 'nice feeling', but early versions
of JSON Schema used URIs with empty fragments as their $ids, and
even though later versions don't, and future versions may even
make this an error, a downstream library (including jsonschema)
is likely to copy these $ids exactly as-is and expect the resource
to be retrievable, despite the last commit making some of this
behavior JSON Schema-specific. We may revisit this all later if/when
it turns out to matter.

Refs: python-jsonschema/jsonschema#1064